requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
aiofiles==24.1.0
    # via fastapi-admin
aioredis==2.0.1
    # via fastapi-admin
aiosqlite==0.21.0
    # via tortoise-orm
alembic==1.15.1
    # via -r requirements.in
annotated-types==0.7.0
    # via pydantic
anyio==4.8.0
    # via
    #   starlette
    #   watchfiles
async-timeout==5.0.1
    # via
    #   aioredis
    #   asyncpg
asyncpg==0.30.0
    # via -r requirements.in
babel==2.17.0
    # via fastapi-admin
bcrypt==4.3.0
    # via
    #   -r requirements.in
    #   fastapi-admin
click==8.1.8
    # via uvicorn
exceptiongroup==1.2.2
    # via anyio
fastapi==0.115.11
    # via
    #   -r requirements.in
    #   fastapi-admin
fastapi-admin==1.0.4
    # via -r requirements.in
greenlet==3.1.1
    # via sqlalchemy
h11==0.14.0
    # via uvicorn
httptools==0.6.4
    # via uvicorn
idna==3.10
    # via anyio
iso8601==2.1.0
    # via tortoise-orm
jinja2==3.1.6
    # via fastapi-admin
mako==1.3.9
    # via alembic
markupsafe==3.0.2
    # via
    #   jinja2
    #   mako
pendulum==3.0.0
    # via fastapi-admin
pydantic==2.10.6
    # via fastapi
pydantic-core==2.27.2
    # via pydantic
pypika-tortoise==0.5.0
    # via tortoise-orm
python-dateutil==2.9.0.post0
    # via
    #   pendulum
    #   time-machine
python-dotenv==1.0.1
    # via
    #   -r requirements.in
    #   uvicorn
python-multipart==0.0.20
    # via fastapi-admin
pytz==2025.1
    # via tortoise-orm
pyyaml==6.0.2
    # via uvicorn
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.38
    # via
    #   -r requirements.in
    #   alembic
starlette==0.46.0
    # via fastapi
time-machine==2.16.0
    # via pendulum
tortoise-orm==0.24.2
    # via fastapi-admin
typing-extensions==4.12.2
    # via
    #   aioredis
    #   aiosqlite
    #   alembic
    #   anyio
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   uvicorn
tzdata==2025.1
    # via pendulum
uvicorn[standard]==0.34.0
    # via fastapi-admin
uvloop==0.21.0
    # via uvicorn
watchfiles==1.0.4
    # via uvicorn
websockets==15.0.1
    # via uvicorn
